BGDIINF_SB-3194: Moved LayerCatalogue out of topics
Because this component will be reused by the import tool move it out of topics.
